requirements.txt
mezzanine==1.0
django-classy-tags
